Effects:

* filtered main submission result chart from ~4 s to ~1 s
* helloworld result chart from ~100 ms to ~10 ms
* no more index_merge queries for `user_attempted_ids` and `user_completed_ids`